Move APIHandler to a package
